requirements.txt
torch>=1.12.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0
huggingface-hub>=0.15.0
tokenizers>=0.13.0
numpy>=1.23.0
pandas>=1.5.0
tqdm>=4.65.0
wandb>=0.15.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.0
nltk>=3.8.0
spacy>=3.5.0
pyyaml>=6.0
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
